# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Riwi Messaging Platform — Thin Backend API",
    "description": "API for Riwi internal messaging platform featuring Smart Database architecture, RLS authorization, STOMP WebSocket and AI Copilot.",
    "contact": {
      "name": "Riwi Co. S.A.S.",
      "email": "dev@riwi.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/copilot/query": {
      "post": {
        "tags": [
          "copilot-controller"
        ],
        "operationId": "queryCopilot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "name": "BearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}